+++ src/main/java/subway/controller/SubwayManager.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.controller;
 /*
  * SubwayManager
  *
