+++ src/main/java/subway/manager/InputSectionManager.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.manager;
 /*
  * InputSectionManager
  *
@@ -8,6 +8,7 @@ package subway;
  *
  * Copyright (c) by Davinci.J
  */
+import subway.controller.SubwayManager;
 import subway.domain.*;
 
 import java.util.Arrays;
